101-mul.c

+++ 0x0C-more_malloc_free/101-mul.c
@@ -10,7 +10,7 @@
 
 int main(int argc, char **argv)
 {
-	long long int mul, num1, num2;
+	long int mul, num1, num2;
 
 	if (argc != 3)
 	{
@@ -22,7 +22,7 @@ int main(int argc, char **argv)
 	num2 = atoi(argv[2]);
 	mul = num1 * num2;
 
-	printf("%lld\n", mul);
+	printf("%ld\n", mul);
 
 	return (0);
 }
